+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added Bancomat Pay
 - Fixed min/max amount validation
+- Added Apple Pay Multibrowser
 
 
 ## [3.2.6] - 2026-07-09